refactor: Remove old safe creation components
