--- src/components/create-safe/status/styles.module.css
@@ -1,10 +1,0 @@
-.image {
-  width: 138px;
-  height: 94px;
-}
-
-.loading {
-  padding: 1px 12px;
-  border-radius: 4px;
-  background-color: #fff;
-}
