--- src/components/create-safe/steps/styles.module.css
@@ -1,51 +1,0 @@
-.connectionOptions {
-  display: flex;
-  flex-direction: row;
-  align-items: flex-start;
-  gap: var(--space-12);
-  padding: var(--space-2) var(--space-6);
-}
-
-.wallet {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  gap: var(--space-3);
-}
-
-.pairing {
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-  gap: var(--space-2);
-}
-
-.pairing > svg {
-  padding-top: 20px;
-}
-
-.pairing > div {
-  display: flex;
-  flex-direction: column;
-  align-items: flex-start;
-  max-width: 240px;
-  gap: var(--space-1);
-}
-
-.pairing > div:first-child {
-  align-items: center;
-}
-
-.pairing :global .MuiTypography-alignCenter {
-  text-align: unset;
-}
-
-@media (max-width: 600px) {
-  .pairing {
-    display: none;
-  }
-
-  .connectionOptions {
-    justify-content: center;
-  }
-}
