+++ src/components/new-safe/OwnerRow/index.tsx
@@ -7,7 +7,7 @@ import DeleteIcon from '@/public/images/common/delete.svg'
 import { useFormContext, useWatch } from 'react-hook-form'
 import { useAddressResolver } from '@/hooks/useAddressResolver'
 import EthHashInfo from '@/components/common/EthHashInfo'
-import type { NamedAddress } from '@/components/create-safe/types'
+import type { NamedAddress } from '@/components/new-safe/create/types'
 import useWallet from '@/hooks/wallets/useWallet'
 import { sameAddress } from '@/utils/addresses'
 import css from './styles.module.css'
