+++ src/components/new-safe/create/steps/Step3/index.tsx
@@ -4,14 +4,14 @@ import ChainIndicator from '@/components/common/ChainIndicator'
 import EthHashInfo from '@/components/common/EthHashInfo'
 import { useCurrentChain } from '@/hooks/useChains'
 import useGasPrice from '@/hooks/useGasPrice'
-import { useEstimateSafeCreationGas } from '@/components/create-safe/useEstimateSafeCreationGas'
+import { useEstimateSafeCreationGas } from '@/components/new-safe/create/useEstimateSafeCreationGas'
 import { formatVisualAmount } from '@/utils/formatters'
 import type { StepRenderProps } from '@/components/new-safe/CardStepper/useCardStepper'
 import type { NewSafeFormData } from '@/components/new-safe/create'
 import css from '@/components/new-safe/create/steps/Step3/styles.module.css'
 import layoutCss from '@/components/new-safe/create/styles.module.css'
 import { getFallbackHandlerContractInstance } from '@/services/contracts/safeContracts'
-import { computeNewSafeAddress } from '@/components/create-safe/logic'
+import { computeNewSafeAddress } from '@/components/new-safe/create/steps/Step4/logic'
 import useWallet from '@/hooks/wallets/useWallet'
 import { useWeb3 } from '@/hooks/wallets/web3'
 import useLocalStorage from '@/services/local-storage/useLocalStorage'
