+++ src/components/new-safe/create/steps/Step4/index.tsx
@@ -10,7 +10,7 @@ import useWallet from '@/hooks/wallets/useWallet'
 import useIsWrongChain from '@/hooks/useIsWrongChain'
 import type { NewSafeFormData } from '@/components/new-safe/create'
 import type { StepRenderProps } from '@/components/new-safe/CardStepper/useCardStepper'
-import type { PendingSafeTx } from '@/components/create-safe/types'
+import type { PendingSafeTx } from '@/components/new-safe/create/types'
 import useSafeCreationEffects from '@/components/new-safe/create/steps/Step4/useSafeCreationEffects'
 import { SafeCreationStatus, useSafeCreation } from '@/components/new-safe/create/steps/Step4/useSafeCreation'
 import StatusStepper from '@/components/new-safe/create/steps/Step4/StatusStepper'
