+++ src/components/new-safe/create/steps/Step4/logic/index.ts
@@ -1,6 +1,10 @@
 import type { Web3Provider, JsonRpcProvider } from '@ethersproject/providers'
 import type { ChainInfo } from '@safe-global/safe-gateway-typescript-sdk'
-import { getProxyFactoryContractInstance } from '@/services/contracts/safeContracts'
+import {
+  getFallbackHandlerContractInstance,
+  getGnosisSafeContractInstance,
+  getProxyFactoryContractInstance,
+} from '@/services/contracts/safeContracts'
 import type { ConnectedWallet } from '@/services/onboard'
 import { BigNumber } from '@ethersproject/bignumber'
 import { SafeCreationStatus } from '@/components/new-safe/create/steps/Step4/useSafeCreation'
@@ -8,7 +12,7 @@ import { didRevert, type EthersError } from '@/utils/ethers-utils'
 import { Errors, logError } from '@/services/exceptions'
 import { ErrorCode } from '@ethersproject/logger'
 import { isWalletRejection } from '@/utils/wallets'
-import type { PendingSafeTx } from '@/components/create-safe/types'
+import type { PendingSafeTx } from '@/components/new-safe/create/types'
 import type { NewSafeFormData } from '@/components/new-safe/create'
 import type { UrlObject } from 'url'
 import chains from '@/config/chains'
@@ -17,7 +21,92 @@ import { SAFE_APPS_EVENTS, trackEvent } from '@/services/analytics'
 import type { AppDispatch, AppThunk } from '@/store'
 import { showNotification } from '@/store/notificationsSlice'
 import { formatError } from '@/hooks/useTxNotifications'
-import { encodeSafeCreationTx } from '@/components/create-safe/logic'
+import { SafeFactory } from '@safe-global/safe-core-sdk'
+import type Safe from '@safe-global/safe-core-sdk'
+import type { DeploySafeProps } from '@safe-global/safe-core-sdk'
+import { createEthersAdapter } from '@/hooks/coreSDK/safeCoreSDK'
+import type { PredictSafeProps } from '@safe-global/safe-core-sdk/dist/src/safeFactory'
+import { getSafeInfo, type SafeInfo } from '@safe-global/safe-gateway-typescript-sdk'
+import { backOff } from 'exponential-backoff'
+import { LATEST_SAFE_VERSION } from '@/config/constants'
+import { EMPTY_DATA, ZERO_ADDRESS } from '@safe-global/safe-core-sdk/dist/src/utils/constants'
+
+export type SafeCreationProps = {
+  owners: string[]
+  threshold: number
+  saltNonce: number
+}
+
+/**
+ * Prepare data for creating a Safe for the Core SDK
+ */
+export const getSafeDeployProps = (
+  safeParams: SafeCreationProps,
+  callback: (txHash: string) => void,
+  chainId: string,
+): PredictSafeProps & { callback: DeploySafeProps['callback'] } => {
+  const fallbackHandler = getFallbackHandlerContractInstance(chainId)
+
+  return {
+    safeAccountConfig: {
+      threshold: safeParams.threshold,
+      owners: safeParams.owners,
+      fallbackHandler: fallbackHandler.getAddress(),
+    },
+    safeDeploymentConfig: {
+      saltNonce: safeParams.saltNonce.toString(),
+    },
+    callback,
+  }
+}
+
+/**
+ * Create a Safe creation transaction via Core SDK and submits it to the wallet
+ */
+export const createNewSafe = async (ethersProvider: Web3Provider, props: DeploySafeProps): Promise<Safe> => {
+  const ethAdapter = createEthersAdapter(ethersProvider)
+
+  const safeFactory = await SafeFactory.create({ ethAdapter })
+  return safeFactory.deploySafe(props)
+}
+
+/**
+ * Compute the new counterfactual Safe address before it is actually created
+ */
+export const computeNewSafeAddress = async (ethersProvider: Web3Provider, props: PredictSafeProps): Promise<string> => {
+  const ethAdapter = createEthersAdapter(ethersProvider)
+
+  const safeFactory = await SafeFactory.create({ ethAdapter })
+  return safeFactory.predictSafeAddress(props)
+}
+
+/**
+ * Encode a Safe creation transaction NOT using the Core SDK because it doesn't support that
+ * This is used for gas estimation.
+ */
+export const encodeSafeCreationTx = ({
+  owners,
+  threshold,
+  saltNonce,
+  chain,
+}: SafeCreationProps & { chain: ChainInfo }) => {
+  const safeContract = getGnosisSafeContractInstance(chain, LATEST_SAFE_VERSION)
+  const proxyContract = getProxyFactoryContractInstance(chain.chainId)
+  const fallbackHandlerContract = getFallbackHandlerContractInstance(chain.chainId)
+
+  const setupData = safeContract.encode('setup', [
+    owners,
+    threshold,
+    ZERO_ADDRESS,
+    EMPTY_DATA,
+    fallbackHandlerContract.getAddress(),
+    ZERO_ADDRESS,
+    '0',
+    ZERO_ADDRESS,
+  ])
+
+  return proxyContract.encode('createProxyWithNonce', [safeContract.getAddress(), setupData, saltNonce])
+}
 
 /**
  * Encode a Safe creation tx in a way that we can store locally and monitor using _waitForTransaction
@@ -45,6 +134,35 @@ export const getSafeCreationTxInfo = async (
     value: BigNumber.from(0),
     startBlock: await provider.getBlockNumber(),
   }
+}
+
+export const estimateSafeCreationGas = async (
+  chain: ChainInfo,
+  provider: JsonRpcProvider,
+  from: string,
+  safeParams: SafeCreationProps,
+): Promise<BigNumber> => {
+  const proxyFactoryContract = getProxyFactoryContractInstance(chain.chainId)
+  const encodedSafeCreationTx = encodeSafeCreationTx({ ...safeParams, chain })
+
+  return provider.estimateGas({
+    from: from,
+    to: proxyFactoryContract.getAddress(),
+    data: encodedSafeCreationTx,
+  })
+}
+
+export const pollSafeInfo = async (chainId: string, safeAddress: string): Promise<SafeInfo> => {
+  // exponential delay between attempts for around 4 min
+  return backOff(() => getSafeInfo(chainId, safeAddress), {
+    startingDelay: 750,
+    maxDelay: 20000,
+    numOfAttempts: 19,
+    retry: (e) => {
+      console.info('waiting for client-gateway to provide safe information', e)
+      return true
+    },
+  })
 }
 
 export const handleSafeCreationError = (error: EthersError) => {
