+++ src/components/new-safe/create/steps/Step4/useSafeCreationEffects.ts
@@ -1,9 +1,9 @@
 import type { Dispatch, SetStateAction } from 'react'
 import { useEffect } from 'react'
-import { pollSafeInfo } from '@/components/create-safe/logic'
+import { pollSafeInfo } from '@/components/new-safe/create/steps/Step4/logic'
 import { SafeCreationStatus } from '@/components/new-safe/create/steps/Step4/useSafeCreation'
 import { CREATE_SAFE_EVENTS, trackEvent } from '@/services/analytics'
-import { updateAddressBook } from '@/components/create-safe/logic/address-book'
+import { updateAddressBook } from '@/components/new-safe/create/steps/Step4/logic/address-book'
 import { useAppDispatch } from '@/store'
 import useChainId from '@/hooks/useChainId'
 import type { PendingSafeData } from '@/components/new-safe/create/steps/Step4/index'
