+++ src/components/new-safe/load/index.tsx
@@ -5,7 +5,7 @@ import { LOAD_SAFE_CATEGORY } from '@/services/analytics'
 import { Container, Grid, Typography } from '@mui/material'
 import { CardStepper } from '@/components/new-safe/CardStepper'
 import type { TxStepperProps } from '@/components/new-safe/CardStepper/useCardStepper'
-import type { NamedAddress } from '@/components/create-safe/types'
+import type { NamedAddress } from '@/components/new-safe/create/types'
 import SetAddressStep from '@/components/new-safe/load/steps/SetAddressStep'
 import { AppRoutes } from '@/config/routes'
 import SafeOwnerStep from '@/components/new-safe/load/steps/SafeOwnerStep'
