+++ src/components/new-safe/load/steps/SafeOwnerStep/index.tsx
@@ -7,7 +7,7 @@ import type { StepRenderProps } from '@/components/new-safe/CardStepper/useCardS
 import type { LoadSafeFormData } from '@/components/new-safe/load'
 import useAsync from '@/hooks/useAsync'
 import useChainId from '@/hooks/useChainId'
-import type { NamedAddress } from '@/components/create-safe/types'
+import type { NamedAddress } from '@/components/new-safe/create/types'
 import layoutCss from '@/components/new-safe/create/styles.module.css'
 import ArrowBackIcon from '@mui/icons-material/ArrowBack'
 import { OwnerRow } from '@/components/new-safe/OwnerRow'
